+++ src/config.ts
@@ -93,6 +93,7 @@ export const DEFAULT_CONFIG: Config = {
     enableIncrementalLoad: true,         // Continue loading in background after initial load
     incrementalLoadDelayMS: 500,         // Delay between incremental loads of toots
     maxTimelineTootsToFetch: 2_500,      // How many standard timeline toots to pull
+    // incrementalLoadDelayMS: 2500,         // Delay between incremental loads of toots
     // maxTimelineTootsToFetch: 800,       // useful dev options for faster load
     maxTimelineHoursToFetch: 168,        // Maximum length of time to pull timeline toots for
     numTootsInFirstFetch: 80,            // How many toots to pull in the first fetch